+++ lib/bloc_helpers.dart
@@ -1,5 +1,5 @@
 library bloc_helpers;
 
 export 'src/bloc.dart';
-export 'src/request.dart';
+export 'src/async.dart';
 export 'src/selector.dart';