+++ lib/src/request.dart
@@ -1,64 +1,59 @@
-import 'dart:async';
-import 'package:rxdart/rxdart.dart';
-import 'package:meta/meta.dart';
-import 'bloc.dart';
+part of 'async.dart';
 
-typedef Future<Response> RequestHandler<Request, Response>(Request input);
+mixin _RequestMixin<Request, Response> on AsyncTaskBloc<Request, Response> {
+  var _markerKey = Object();
 
-/// Helper class to implement asynchronous call to a server
-/// Need to implement the [request] method
-abstract class RequestBloc<Request, Response> extends Bloc {
-  var _markerKey = new Object();
-
-  final _requestPublisher = new PublishSubject<Request>();
-
-  // ignore: close_sinks
-  final _responsePublisher = new PublishSubject<Response>();
-
-  // ignore: close_sinks
-  final _loadingBehavior = new BehaviorSubject<bool>(seedValue: false);
-
-  RequestBloc() {
-    onRequest.listen(_handleRequest);
-  }
-
-  Future<void> _handleRequest(Request input) async {
-    _loadingBehavior.add(true);
-
+  Future<void> _doRequest(Request input) async {
+    _runningBehavior.add(true);
     try {
-      final key = _markerKey = new Object();
-      final res = await request(input);
+      final key = _markerKey = Object();
+
+      final res = await run(input);
 
       if (key != _markerKey) return;
 
-      _responsePublisher.add(res);
+      _resultPublisher.add(res);
     } catch (e, s) {
-      _responsePublisher.addError(e, s);
+      _resultPublisher.addError(e, s);
+    } finally {
+      _runningBehavior.add(false);
     }
-    _loadingBehavior.add(false);
-  }
-
-  @mustCallSuper
-  @override
-  FutureOr<void> dispose() async {
-    await _requestPublisher.close();
-
-    super.dispose();
   }
 
   /// Sink to trigger the request
-  /// the response and errors are push in the [onResponse] stream
-  Sink<Request> get requestSink => _requestPublisher.sink;
+  /// the response and errors are push in the [onResult] stream
+  @Deprecated('Use callSink')
+  Sink<Request> get requestSink => callSink;
 
   /// Stream representing the current state of the bloc
   /// true if a request is ongoing
-  ValueObservable<bool> get onLoading => _loadingBehavior.stream;
+  @Deprecated('Use onRunning')
+  ValueObservable<bool> get onLoading => onRunning;
 
   /// Request stream
-  Observable<Request> get onRequest => _requestPublisher.stream;
+  @Deprecated('Use onCall')
+  Observable<Request> get onRequest => onCall;
 
   /// Response stream
-  Observable<Response> get onResponse => _responsePublisher.stream;
+  @Deprecated('Use onResult')
+  Observable<Response> get onResponse => onResult;
+}
+
+/// Helper class to implement asynchronous call to a server
+/// Need to implement the [request] method
+///
+/// The difference with an AsyncTaskBloc
+/// Is that the taskHandler is using a key to mark the request
+/// if you trigger 2 request at the same time of at a small interval
+/// ```dart
+/// callSink.add(request);
+/// callSink.add(request);
+/// ```
+/// The result of the first one will be ignored
+abstract class RequestBloc<Request, Response>
+    extends AsyncTaskBloc<Request, Response>
+    with _RequestMixin<Request, Response> {
+  RequestBloc();
 
   /// Create a Request Bloc by passing the request function
   ///
@@ -69,11 +64,17 @@ abstract class RequestBloc<Request, Response> extends Bloc {
   ///   ...
   /// }
   /// ```
-  factory RequestBloc.func(RequestHandler<Request, Response> handler) =>
+  factory RequestBloc.func(TaskHandler<Request, Response> handler) =>
       new _RequestBloc<Request, Response>(handler);
 
   @protected
-  Future<Response> request(Request input);
+  FutureOr<Response> request(Request input);
+
+  @protected
+  FutureOr<Response> run(Request input) => request(input);
+
+  @override
+  Future<void> _handleTask(Request input) => _doRequest(input);
 }
 
 /// Helper class to implement asynchronous call to a server
@@ -82,67 +83,18 @@ abstract class RequestBloc<Request, Response> extends Bloc {
 ///
 /// ```dart
 /// cachedRequestBloc.requestSink.add('foo'); /// will call [request]
-/// cachedRequestBloc.onResponse.first; /// response 'a'
+/// cachedRequestBloc.onResult.first; /// response 'a'
 ///
 /// cachedRequestBloc.requestSink.add('foo'); /// same input, won't call [request]
-/// cachedRequestBloc.onResponse.first; /// response 'a'
+/// cachedRequestBloc.onResult.first; /// response 'a'
 /// ```
 ///
 /// If the request input change it will invalidate the cache and call [request]
 abstract class CachedRequestBloc<Request, Response>
-    extends RequestBloc<Request, Response> {
-  var _cached = false;
-
-  // ignore: close_sinks
-  final BehaviorSubject<Response> _cachedResponseBehavior;
-
-  // ignore: close_sinks
-  final _cachedRequestBehavior = new BehaviorSubject<Request>();
-  final _invalidatePublisher = new PublishSubject<Response>();
-
-  /// [seedValue] will init the value of the [cachedResponse]
-  CachedRequestBloc({Response seedValue})
-      : _cachedResponseBehavior =
-            new BehaviorSubject<Response>(seedValue: seedValue),
-        super() {
-    onResponse.listen(_onResponse, onError: _onError);
-    _invalidatePublisher.stream.listen((value) {
-      _cached = false;
-      _cachedRequestBehavior.add(null);
-
-      if (disposed) return;
-      _cachedResponseBehavior.add(value ?? seedValue);
-    });
-  }
-
-  @override
-  Future<void> _handleRequest(Request input) async {
-    if (_cachedRequestBehavior.value == input && _cached) {
-      _responsePublisher.add(_cachedResponseBehavior.value);
-      return;
-    }
-    _cachedRequestBehavior.add(input);
-    super._handleRequest(input);
-  }
-
-  void _onError(e, s) {
-    if (disposed) return;
-    _cachedResponseBehavior.addError(e, s);
-  }
-
-  void _onResponse(Response response) {
-    _cached = true;
-    if (disposed) return;
-    _cachedResponseBehavior.add(response);
-  }
-
-  @override
-  @mustCallSuper
-  FutureOr<void> dispose() async {
-    await _invalidatePublisher.close();
-    await _cachedResponseBehavior.close();
-    super.dispose();
-  }
+    extends AsyncCachedTaskBloc<Request, Response>
+    with _RequestMixin<Request, Response> {
+  /// [seedValue] will init the value of the [cachedResult]
+  CachedRequestBloc({Response seedValue}) : super(seedValue: seedValue);
 
   /// Create a Cached Request Bloc by passing the request function
   ///
@@ -153,36 +105,49 @@ abstract class CachedRequestBloc<Request, Response>
   ///   ...
   /// }
   /// ```
-  factory CachedRequestBloc.func(RequestHandler<Request, Response> handler) =>
+  factory CachedRequestBloc.func(TaskHandler<Request, Response> handler) =>
       new _CachedRequestBloc<Request, Response>(handler);
 
   /// cached response stream
   /// Use a Behavior subject so will emit the last value at each `listen`
-  ValueObservable<Response> get cachedResponse => _cachedResponseBehavior.stream;
-
-  /// Sink to invalidate the cache
-  /// Can take a value if you want to put back the seedValue of the cachedResponse
-  Sink<Response> get invalidateCacheSink => _invalidatePublisher.sink;
+  @Deprecated('Use cachedResult')
+  ValueObservable<Response> get cachedResponse => cachedResult;
 
   /// Sink to manualy update the cachedResponse
-  Sink<Response> get updateCachedResponseSink => _cachedResponseBehavior.sink;
+  @Deprecated('Use updateCachedResultSink')
+  Sink<Response> get updateCachedResponseSink => updateCachedResultSink;
+
+  @protected
+  FutureOr<Response> request(Request input);
+
+  @protected
+  FutureOr<Response> run(Request input) => request(input);
+
+  @override
+  Future<void> _handleTask(Request input) async {
+    final hitCache = _handleCache(input);
+
+    if (hitCache) return;
+
+    await _doRequest(input);
+  }
 }
 
 class _CachedRequestBloc<Request, Response>
     extends CachedRequestBloc<Request, Response> {
-  final RequestHandler<Request, Response> _request;
+  final TaskHandler<Request, Response> _request;
 
   _CachedRequestBloc(this._request) : super();
 
   @override
-  Future<Response> request(Request input) => _request(input);
+  FutureOr<Response> request(Request input) => _request(input);
 }
 
 class _RequestBloc<Request, Response> extends RequestBloc<Request, Response> {
-  final RequestHandler<Request, Response> _request;
+  final TaskHandler<Request, Response> _request;
 
   _RequestBloc(this._request) : super();
 
   @override
-  Future<Response> request(Request input) => _request(input);
+  FutureOr<Response> request(Request input) => _request(input);
 }